Customized admin.py
Simply changed embed color instances from author.id to a specific color 0x6edfba for personal use. Hoping to later come back to this so that it can be changed to any hex color requested, so that it benefits vrt and the entire community. 

Not even entirely sure i've updated it correctly yet anyhow...